Logout & mock-oauth-flow Moved logout to its own file to deal with MSW & custom mocking conflicting. mock-http-server changes reflect the mockGrantAuthorization used from mockOAuthServerCallback
