+++ packages/wrangler/src/__tests__/user.test.ts
@@ -1,13 +1,9 @@
-import fs from "node:fs";
-import path from "node:path";
 import { rest } from "msw";
-import { getGlobalWranglerConfigPath } from "../global-wrangler-config-path";
 import { CI } from "../is-ci";
 import {
 	loginOrRefreshIfRequired,
 	readAuthConfigFile,
 	requireAuth,
-	USER_AUTH_CONFIG_FILE,
 	writeAuthConfigFile,
 } from "../user";
 import { mockConsoleMethods } from "./helpers/mock-console";
@@ -28,7 +24,7 @@ describe("User", () => {
 	runInTempDir();
 	const std = mockConsoleMethods();
 	// TODO: Implement these two mocks with MSW
-	const { mockOAuthServerCallback, mockGrantAuthorization } = mockOAuthFlow();
+	const { mockOAuthServerCallback } = mockOAuthFlow();
 	const { setIsTTY } = useMockIsTTY();
 
 	beforeEach(() => {
@@ -38,13 +34,12 @@ describe("User", () => {
 
 	describe("login", () => {
 		it("should login a user when `wrangler login` is run", async () => {
-			mockOAuthServerCallback();
-			mockGrantAuthorization({ respondWith: "success" });
+			mockOAuthServerCallback("success");
+
 			let counter = 0;
 			msw.use(
-				rest.post("*/oauth2/token", (_, response, context) => {
-					counter++;
-					expect(counter).toBe(1);
+				rest.post("*/oauth2/token", async (_, response, context) => {
+					counter += 1;
 
 					return response.once(
 						context.status(200),
@@ -57,14 +52,15 @@ describe("User", () => {
 					);
 				})
 			);
+
 			await runWrangler("login");
 
+			expect(counter).toBe(1);
 			expect(std.out).toMatchInlineSnapshot(`
 			"Attempting to login via OAuth...
 			Opening a link in your default browser: https://dash.cloudflare.com/oauth2/auth?response_type=code&client_id=54d11594-84e4-41aa-b438-e81b8fa78ee7&redirect_uri=http%3A%2F%2Flocalhost%3A8976%2Foauth%2Fcallback&scope=account%3Aread%20user%3Aread%20workers%3Awrite%20workers_kv%3Awrite%20workers_routes%3Awrite%20workers_scripts%3Awrite%20workers_tail%3Aread%20d1%3Awrite%20pages%3Awrite%20zone%3Aread%20offline_access&state=MOCK_STATE_PARAM&code_challenge=MOCK_CODE_CHALLENGE&code_challenge_method=S256
 			Successfully logged in."
 		`);
-
 			expect(readAuthConfigFile()).toEqual<UserAuthConfig>({
 				api_token: undefined,
 				oauth_token: "test-access-token",
@@ -75,46 +71,25 @@ describe("User", () => {
 		});
 	});
 
-	describe("logout", () => {
-		it("should exit with a message stating the user is not logged in", async () => {
-			await runWrangler("logout");
-			expect(std.out).toMatchInlineSnapshot(`"Not logged in, exiting..."`);
-		});
-
-		it("should logout user that has been properly logged in", async () => {
-			writeAuthConfigFile({
-				oauth_token: "some-oauth-tok",
-				refresh_token: "some-refresh-tok",
-			});
-			let counter = 0;
-			msw.use(
-				rest.post("*/oauth2/token/revoke", (_, response, context) => {
-					counter++;
-					// Make sure that we made the request to logout.
-					expect(counter).toBe(1);
-					response.once(context.status(200), context.text(""));
-				})
-			);
-
-			await runWrangler("logout");
-			expect(std.out).toMatchInlineSnapshot(`"Successfully logged out."`);
-
-			// Make sure that logout removed the config file containing the auth tokens.
-			const config = path.join(
-				getGlobalWranglerConfigPath(),
-				USER_AUTH_CONFIG_FILE
-			);
-			expect(fs.existsSync(config)).toBeFalsy();
-		});
-	});
-
-	it("should handle errors for failed token refresh", async () => {
+	it("should handle errors for failed token refresh in a non-interactive environment", async () => {
 		setIsTTY(false);
 		writeAuthConfigFile({
 			oauth_token: "hunter2",
 			refresh_token: "Order 66",
 		});
 		// TODO: Use MSW to handle `/token` endpoints from different calls
+		let counter = 0;
+		msw.use(
+			rest.post("*/oauth2/token", async (request, response, context) => {
+				counter += 1;
+				return response.once(
+					context.status(400),
+					context.body(
+						`<html> <body> This shouldn't be sent, but should be handled </body> </html>`
+					)
+				);
+			})
+		);
 
 		// Handles the requireAuth error throw from failed login that is unhandled due to directly calling it here
 		await expect(
@@ -122,6 +97,7 @@ describe("User", () => {
 		).rejects.toThrowErrorMatchingInlineSnapshot(
 			`"In a non-interactive environment, it's necessary to set a CLOUDFLARE_API_TOKEN environment variable for wrangler to work. Please go to https://developers.cloudflare.com/api/tokens/create/ for instructions on how to create an api token, and assign its value to CLOUDFLARE_API_TOKEN."`
 		);
+		expect(counter).toBe(0);
 	});
 
 	it("should confirm no error message when refresh is successful", async () => {
